added fake data to temp database

+++ Domain/Models/IngredientRecipe.cs
@@ -7,5 +7,5 @@ public class IngredientRecipe
     public int IngredientId { get; set; }
     public int RecipeId { get; set; }
     public UnitOfMeasurement UnitOfMeasurement { get; set; }
-    public int Quantity { get; set; }
+    public double Quantity { get; set; }
 }
